feat: alter type of column "weight" from int to float

+++ src/infra/entities/item.entity.ts
@@ -23,7 +23,7 @@ export class Item {
   @Column('varchar', { default: '', nullable: false })
   description: string;
 
-  @Column('int', { default: 0 })
+  @Column('float', { default: 0 })
   weight: number;
 
   @Column('varchar', { default: '', nullable: false })
